pyproject.toml
[project]
name = "langchain-mcp-tools"
version = "0.1.11"
description = "Model Context Protocol (MCP) To LangChain Tools Conversion Utility"
keywords = [
    "modelcontextprotocol",
    "mcp",
    "mcp-client",
    "langchain",
    "langchain-python",
    "tool-call",
    "tool-calling",
    "python",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema-pydantic>=0.6",
    "langchain>=0.3.14",
    "mcp>=1.6.0",
    "pyjson5>=1.6.8",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
dev = [
    "dotenv>=0.9.9",
    "langchain-anthropic>=0.3.1",
    "langchain-groq>=0.2.3",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
langchain_mcp_tools = ["py.typed"]

[project.urls]
"Bug Tracker" = "https://github.com/hideya/langchain-mcp-tools-py/issues"
"Source Code" = "https://github.com/hideya/langchain-mcp-tools-py"
